+++ insanic/log/formatters.py
@@ -1,4 +1,5 @@
 import aiotask_context
+import inspect
 import logging
 import time
 import socket
@@ -121,6 +122,8 @@ class JSONFormatter(logging.Formatter):
             except ValueError:
                 msg = record.getMessage()
                 self._add_dic(data, {'message': msg})
+        elif isinstance(msg, Exception) or inspect.isclass(msg):
+            self._add_dic(data, {"message": str(msg)})
         else:
             self._add_dic(data, {'message': msg})