+++ CHANGELOG.md
@@ -2,7 +2,25 @@ Changelog for insanic
 =====================
 
 
-0.7.11 (unreleased)
+0.7.12 (unreleased)
+-------------------
+
+- BREAKING: `request_timeout` arguments of dispatch methods have been changed to
+`response_timeout`.
+- BREAKING: If the response has timed out, Insanic will raise `503 Service Unavailable` with message `{service_name} has timed out.`,
+not `408 Request Timeout` anymore.
+- FIX: All timeout variables will be used as `ClientTimeout` class, not primitive integer.
+See [https://docs.aiohttp.org/en/stable/client_quickstart.html?#timeouts](https://docs.aiohttp.org/en/stable/client_quickstart.html?#timeouts)
+- UPGRADE: minimum requirement for aiohttp updated from 3.1.0 -> 3.3.0 
+- UPGRADE: increase log level to critical for uncaught exceptions
+- CHORE: changed ttl_dns_cache on service session to reflect dns cache ttl of route53
+- UPDATE: default response timeout for inter-service to 5s
+- UPDATE: added default connection timeout to 1s
+- CHORE: updated config to include sanic's default configs
+- FIX: logger now sets value as exception message
+
+
+0.7.11 (2019-05-29)
 -------------------
 
 - BREAKING: THROTTLE_NUM_PROXIES deprecated in favor of PROXIES_COUNT